Add a NApp repo information builder.
This script will loop on a NApp '/repo' repository and print on the
stdout the kytos.json for all founded napps ('.napp' files on the tree).
The full output will be a json with one key named "napps" and the value
will be a dictionary with the 'napp-name' (including its version, e.g.:
'author/napp-timestamp-1.napp') as key and the kytos.json value as
value.